OHI/O Hackathon Project 2015
This was a project created in 24 hours for the OHI/O hackathon. I worked
on a team of three creating a Suicide Prevention Simulator application
to spread awareness about suicide and ways to prevent a friend or family
member from taking their life through peaceful conversation. The
application allowed the user to have a text based conversation with the
computer using prewritten XML conversations that could be read in and
run using the application.